changed react-i18next to global mock
